% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal_extract_helpers.R
\name{cal_extract_pressure_data}
\alias{cal_extract_pressure_data}
\title{Extract Pressure Data from Calibration Markup}
\usage{
cal_extract_pressure_data(div)
}
\arguments{
\item{div}{A read-in HTML div element containing calibration data tables to be parsed.}
}
\description{
Extract Pressure Data from Calibration Markup
}
