% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcw.qaqc-package.R
\docType{package}
\name{fcw.qaqc}
\alias{ross.wq.tools}
\alias{ross.wq.tools-package}
\alias{fcw.qaqc}
\title{FCWQAQC: Quality Assurance and Quality Control for Fort Collins Poudre Sonde Network Data}
\description{
Tools for processing, flagging, and analyzing water quality data
from the Fort Collins Watershed monitoring network. This package provides
functions for data validation, quality control flagging, and summary statistics.
}
\author{
\strong{Maintainer}: Juan De La Torre \email{Juan.DeLaTorre@colostate.edu}

Authors:
\itemize{
  \item Kathryn Willi \email{Kathryn.Willi@colostate.edu}
  \item Samuel Struthers \email{Samuel.Struthers@colostate.edu}
}

}
\keyword{internal}
