% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_site_order.R
\name{load_site_order}
\alias{load_site_order}
\title{Load site order definitions from a YAML file}
\usage{
load_site_order(file_path)
}
\arguments{
\item{file_path}{Path to a YAML or CSV file containing site relationships.}
}
\value{
A named list where each element corresponds to a network/system and contains
a character vector of that network's ordered sites from upstream to downstream.
}
\description{
Reads a YAML or csv file that defines upstream and downstream relationships
for each site and returns a named list of site order vectors. If the site is in an independent network,
the name of the list should be the name of that site to avoid confusion/duplicate searches.

For each network, this function constructs a vector containing the upstream sites
(if defined), the site itself, and the downstream sites (if defined).
}
\section{Example YAML file format (see template in inst/templates/template_site_orders.yaml):}{
network:
clp:
joei, cbri, chd, pfal, pbr, pman, pbd, bellvue, salyer, udall, riverbend, cottonwood, elc, archery, riverbluffs
springcreek:
riverbend, springcreek, cottonwood
boxcreek:
elc, boxcreek, archery
sfm:
sfm
mtn_campus:
mtn_campus
}

\section{Example CSV format (see template in inst/templates/template_site_orders.csv):}{
network_name,network
clp,"joei, cbri, chd, pfal, pbr, pman, pbd, bellvue, salyer, udall, riverbend, cottonwood, elc, archery, riverbluffs"
springcreek,"riverbend, springcreek, cottonwood"
boxcreek,"elc, boxcreek, archery"
sfm,sfm
mtn_campus,mtn_campus

Expected Outputs:
site_order_list <- load_site_order_yaml("site_orders.yaml")
site_order_list$clp
#> [1] "joei", "cbri", "chd", "pfal", "pbr", "pman", "pbd", "bellvue", "salyer", "udall", "riverbend", "cottonwood", "elc", "archery", "riverbluffs"

site_order_list$sfm
#> [1] "sfm"
}

\seealso{
\code{\link[=network_check]{network_check()}}
}
